+++ jexxa-core/src/main/java/io/jexxa/infrastructure/drivenadapterstrategy/messaging/JMSQueue.java
@@ -1,0 +1,21 @@
+package io.jexxa.infrastructure.drivenadapterstrategy.messaging;
+
+public class JMSQueue
+{
+    private final String destination;
+
+    JMSQueue(String destination)
+    {
+        this.destination = destination;
+    }
+
+    String getDestination()
+    {
+        return destination;
+    }
+
+    static JMSQueue of(String destination)
+    {
+        return new JMSQueue(destination);
+    }
+}