+++ jexxa-core/src/main/java/io/jexxa/infrastructure/drivenadapterstrategy/messaging/JMSTopic.java
@@ -1,0 +1,21 @@
+package io.jexxa.infrastructure.drivenadapterstrategy.messaging;
+
+public class JMSTopic
+{
+    private final String destination;
+
+    JMSTopic(String destination)
+    {
+        this.destination = destination;
+    }
+
+    String getDestination()
+    {
+        return destination;
+    }
+
+    static JMSTopic of(String destination)
+    {
+        return new JMSTopic(destination);
+    }
+}