+++ CHANGELOG.md
@@ -4,8 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## \[2.x.x] - yyyy-mm-dd
+### Added                                       
+-   Added new fluent API for sending JMS messages  
+      
+### Fixed
+
+### Changed
+-   Old JMS API is declared as deprecated
+
 ## \[2.1.1] - 2020-07-18
-### Added
+### Added                                       
 -   Added tutorials to show usage of Jexxa      
 
 ### Fixed